refactor : 날짜 계산 함수 분리
